Xbox One buttons
Tried playing around with replicating the button selection process on
the xbox one interface as of this commit. To describe the action:
mousing over the button introduces a transparent border (faked with a
colored border, bc you can’t use “transparent” as a style). Thought
about scaling the button as well, as happens in the xbox interface.
Scrapped the whole thing because it looked kind of choppy, and changes
the size of the element in an unpleasant way.